Added pattern test case for 3 more detectors
